Add files via upload
A Labyrinth is simply an m × n grid containing different types of tiles. In our version, there are 3 types of
tiles possible:
• Empty tile. As the name suggests there is nothing on this tile.
• Coin tile. This tile contains a coin. After the coin is collected, this tile will turn into an empty tile.
• Wall tile. This tile contains a wall and it is not possible to visit this tile. This tile will be converted
to an empty tile if the wall is destroyed.
The player will start at the top left corner of the Labyrinth and the goal is to reach the bottom right corner
while collecting as many coins as possible. In one move, the player can either move one tile or perform a
spell. The player can possibly move in 4 directions, up, down, left or right. One of three things can happen:
• The player moved into an empty tile.
• The player moved into a tile with a coin. The coin will be collected and the tile will become empty.
• The player moved into a wall. In this case, a collision will happen and the player will lose a health
point.
The player may also perform one of 2 types of spells:
• Revelio. This spell will be performed on a target tile with a tile type in mind. All the surrounding
tiles (consider the 3 × 3 square around the target) with the desired tile type (coin/wall) will be
revealed.
• Bombarda. This spell will destroy all objects on 3 tiles of the player’s choice, converting these tiles to
empty tiles.
Initially, the player has three health points and can perform a total of three spells